add or deduct virtual points based on the winner functionality is added

+++ src/main/java/com/tekusource/sabongpro/controller/AdminController.java
@@ -108,7 +108,7 @@ public class AdminController extends AbstractController {
 	@RequestMapping(value="/user/allow/access", method=RequestMethod.POST)
 	@ResponseBody
 	public Map<String, ? extends Object> userAllowAccess(@RequestParam("userId") Long userId, 
-														 @RequestParam("streamingAccess") Integer streamingAccess, @RequestParam("virtualPoints") Long virtualPoints){
+														 @RequestParam("streamingAccess") Integer streamingAccess, @RequestParam("virtualPoints") double virtualPoints){
 		Map<String, Object> map = new HashMap<String, Object>();
 		try {
 			User user = (User) userService.getUserBy(userId);
